feat: sync up with internal, 0.59.9

+++ build.sh
@@ -5,7 +5,7 @@ set -e
 
 SCRIPTPATH=$( cd "$(dirname "$0")" ; pwd -P ) 
 OUT_BUILD_DIR="$SCRIPTPATH/build"
-REACT_ROOT="$2"
+PROJECT_ROOT="$2"
 SCRIPTNAME="$0"
 
 # Parameters:
@@ -15,6 +15,8 @@ print_usage () {
 	local _nounderline=`tput rmul`
 	echo "USAGE"
 	echo "$0 -build ${_underline}reactnodemodulepath${_nounderline}"
+	echo ""
+	echo "$0 -build-deps ${_underline}hydranodemodulepath${_nounderline}"
 	echo ""
 	echo "$0 -clean"
 	echo ""
@@ -32,26 +34,38 @@ REACT_LIBS=(
 	Text
 	Vibration
 	WebSocket
+	PushNotificationIOS
+)
+
+# Update these to include the community libs you use and want to precompile
+REACT_COMMUNITY_LIBS=(
+	react-native-webview
+	react-native-gesture-handler
+	@react-native-community/netinfo
 )
 
 REJECT_ARCHIVES=(
 	libdouble-conversion.a
-	libjschelpers.a
+	libjsi.a
+	libjsiexecutor.a
 	libthird-party.a
 	libcxxreact.a
 	libyoga.a
+	libfishhook.a
+	libjsinspector.a
 )
 
 # Parameters:
 # 1. debug/release
 # 2. simulator/device
 # 3. Xcodeproj
+# 4. Should override includes
 build () {
 	case "$1" in
 		debug)
 			local configuration="Debug"
 			;;
-		
+
 		release)
 			local configuration="Release"
 			;;
@@ -62,13 +76,13 @@ build () {
 
 	case "$2" in
 		simulator)
-			local archs="x86_64"
-			local sdk="iphonesimulator10.3"
+			local archs="i386 x86_64"
+			local sdk="iphonesimulator"
 			;;
 
 		device)
 			local archs="armv7 arm64"
-			local sdk="iphoneos10.3"
+			local sdk="iphoneos"
 			;;
 
 		*)
@@ -76,8 +90,14 @@ build () {
 	esac
 	local xcodeproj="$3"
 	local target=$(basename "$xcodeproj" .xcodeproj)
+	local header_search_paths='$(inherited)'
+	if [ "$4" = true ] ; then
+		header_search_paths=\"$SCRIPTPATH/include\"
+	fi
+
 
 	local xcargs=(-project "$xcodeproj"
+		-UseModernBuildSystem=NO
 		-target "$target"
 		-sdk "$sdk"
 		-configuration "$configuration"
@@ -90,6 +110,9 @@ build () {
 		OBJROOT="$OUT_BUILD_DIR/Intermediates"
 		SHARED_PRECOMPS_DIR="$OUT_BUILD_DIR/Intermediates/PrecompiledHeaders"
 		SYMROOT="$OUT_BUILD_DIR/Products"
+		OTHER_CFLAGS='$(inherited) -fembed-bitcode'
+		HEADER_SEARCH_PATHS="$header_search_paths"
+		GCC_GENERATE_DEBUGGING_SYMBOLS=no
 	)
 	echo "xcodebuild ${xcargs[@]}"
 	xcodebuild "${xcargs[@]}" > "$OUT_BUILD_DIR/xcodebuild.$configuration.$2.log"
@@ -97,12 +120,14 @@ build () {
 
 # Parameters
 # 1. Xcodeproj
+# 2. Needs includes override
 build_all_variants () {
 	local xcodeproj="$1"
-	build "debug" "simulator" "$xcodeproj"
-	build "release" "simulator" "$xcodeproj"
-	build "debug" "device" "$xcodeproj"
-	build "release" "device" "$xcodeproj"
+	local includes="$2"
+	build "debug" "simulator" "$xcodeproj" "$includes"
+	build "release" "simulator" "$xcodeproj" "$includes"
+	build "debug" "device" "$xcodeproj" "$includes"
+	build "release" "device" "$xcodeproj" "$includes"
 }
 
 # Parameters:
@@ -113,7 +138,7 @@ join_libraries () {
 		debug)
 			local configuration="Debug"
 			;;
-		
+
 		release)
 			local configuration="Release"
 			;;
@@ -138,28 +163,35 @@ join_libraries () {
 
 	# Delete archives that we don't need
 	for reject_lib in "${REJECT_ARCHIVES[@]}"; do
-		echo "Deleting $build_dir/${reject_lib}"
-		rm "$build_dir/${reject_lib}"
+		local reject_lib_path="$build_dir/${reject_lib}"
+		if [ -f "$reject_lib_path" ]; then
+			 echo "Deleting $reject_lib_path"
+   			 rm "$build_dir/${reject_lib}"
+		fi
 	done
+
+	local universal_binary_name="$3"
 
 	# Link all of React Native's .a files into a single .a file:
 	pushd "$build_dir" > /dev/null
 	echo "$PWD"
-	echo "libtool -static -o \"libReact.$configuration.a\" *.a"
-	find . -name '*.a' -depth 1 -print0 | xargs -0 libtool -static -o "libReact.$1.a" >> "$OUT_BUILD_DIR/xcodebuild.$configuration.$dest.log" 2>&1
+	echo "libtool -static -o \"$universal_binary_name.$configuration.a\" *.a"
+	find . -name '*.a' -depth 1 -print0 | xargs -0 libtool -static -o "$universal_binary_name.$1.a" >> "$OUT_BUILD_DIR/xcodebuild.$configuration.$dest.log" 2>&1
 	popd > /dev/null
 }
 
 # Parameters
 # 1. debug/release
+# 2. ex. libReact/libReactCommunity
 pack_universal_binary () {
 	local configuration="$1"
+	local universal_binary_name="$2"
 	case "$1" in
 		debug)
 			local device_build_dir="$OUT_BUILD_DIR/Products/Debug-iphonesimulator"
 			local sim_build_dir="$OUT_BUILD_DIR/Products/Debug-iphoneos"
 			;;
-		
+
 		release)
 			local device_build_dir="$OUT_BUILD_DIR/Products/Release-iphonesimulator"
 			local sim_build_dir="$OUT_BUILD_DIR/Products/Release-iphoneos"
@@ -168,45 +200,74 @@ pack_universal_binary () {
 		*)
 			print_usage "$SCRIPTNAME"
 	esac
-	
-	join_libraries "$configuration" "simulator"
-	join_libraries "$configuration" "device"	
 
-	# Join the device / simulator .a files into a fat .a file:
+	join_libraries "$configuration" "simulator" "$universal_binary_name"
+	join_libraries "$configuration" "device" "$universal_binary_name"
+
+	# Join the device / simular .a files into a fat .a file:
 	local xcrun_args=(
-		-sdk iphoneos 
-		lipo -create 
-		-output "$OUT_BUILD_DIR/libReact.$configuration.a"
-		"$device_build_dir/libReact.$configuration.a"
-		"$sim_build_dir/libReact.$configuration.a"
+		-sdk iphoneos
+		lipo -create
+		-output "$OUT_BUILD_DIR/$universal_binary_name.a"
+		"$device_build_dir/$universal_binary_name.$configuration.a"
+		"$sim_build_dir/$universal_binary_name.$configuration.a"
 	)
 	echo "xcrun ${xcrun_args[@]}"
 	xcrun "${xcrun_args[@]}"
+	strip -Sx "$OUT_BUILD_DIR/$universal_binary_name.a"
+	pushd "$OUT_BUILD_DIR"
+	zip "$universal_binary_name.$configuration.a.zip" "$universal_binary_name.a"
+	rm "$universal_binary_name.a"
+	popd
 }
 
 case "$1" in
 	-build)
 		mkdir -p "$OUT_BUILD_DIR"
 
-		if [ -z "$REACT_ROOT" ] || [ ! -e "$REACT_ROOT" ] ; then
+		if [ -z "$PROJECT_ROOT" ] || [ ! -e "$PROJECT_ROOT" ] ; then
 			print_usage $0
 		fi
 
-		build_all_variants "$REACT_ROOT/React/React.xcodeproj"
+		build_all_variants "$PROJECT_ROOT/React/React.xcodeproj" false
 
 		for react_lib in "${REACT_LIBS[@]}"; do
-			libdir="$REACT_ROOT/Libraries/$react_lib"
+			libdir="$PROJECT_ROOT/Libraries/$react_lib"
 			xcodeproj=$(find "$libdir" -name '*.xcodeproj' | head -n 1)
-			build_all_variants "$xcodeproj"
+			build_all_variants "$xcodeproj" false
 		done
 
-		pack_universal_binary "debug"
-		pack_universal_binary "release"
+		pack_universal_binary "debug" "libReact"
+		pack_universal_binary "release" "libReact"
 		rm -rf libReact.release.a libReact.debug.a include
-		cp "$OUT_BUILD_DIR/libReact.debug.a" "$OUT_BUILD_DIR/libReact.release.a" .
+		cp "$OUT_BUILD_DIR/libReact.debug.a.zip" "$OUT_BUILD_DIR/libReact.release.a.zip" .
 		cp -R "$OUT_BUILD_DIR/Products/Release-iphoneos/include" .
 		;;
+	
+	-build-deps)
+		if [ ! -d "$SCRIPTPATH/include" ]; then
+			# Fail early if the react native headers are not available in React Native folder.
+  			print_usage "$SCRIPTNAME"
+		fi
 
+		if [ -z "$PROJECT_ROOT" ] || [ ! -e "$PROJECT_ROOT" ] ; then
+			print_usage $0
+		fi
+
+		mkdir -p "$OUT_BUILD_DIR"
+		
+		for react_lib in "${REACT_COMMUNITY_LIBS[@]}"; do
+			libdir="$PROJECT_ROOT/$react_lib"
+			xcodeproj=$(find "$libdir" -name '*.xcodeproj' | head -n 1)
+			build_all_variants "$xcodeproj" true
+		done
+		
+		pack_universal_binary "debug" "libReactCommunity"
+		pack_universal_binary "release" "libReactCommunity"
+		rm -rf libReactCommunity.release.a libReactCommunity.debug.a
+		cp "$OUT_BUILD_DIR/libReactCommunity.debug.a.zip" "$OUT_BUILD_DIR/libReactCommunity.release.a.zip" .
+		;;
+	
 	-clean)
 		rm -rf "$OUT_BUILD_DIR"
 		;;
@@ -214,4 +275,3 @@ case "$1" in
 	*)
 		print_usage "$0"
 esac
-
